ci(experimental/x32): revert
Debian x32 port, the only sane distro that supports the x32 ABI, has
broken packages when used in a multiarch configuration with the usual
x64 ABI, because some versions of transitive dependencies are different
between the x32 port and amd64 repositories, and some packages have
restrictive conflictive requirements that do not allow two slightly
different versions to coexist on the same system.

The proper solution would be to ditch the amd64 repositories and just
use the x32 ones, but the x32 ABI is disabled by default on most
Debian-based kernels, and I bet that GH Actions runners do not have the
necessary kernel boot parameter to enable it back. QEMU does not seem to
have support for the x32 ABI, and we don't want to do heavy hackery on
getting conflicting packages working fine, so let's give up.