tweak(png_file): add dithering level option, improve pass sel. algorithm
Mergu#0001 reached out over Discord to let us know about a ~800 KiB
texture that PackSquash ended up storing on >6 MiB on the ZIP, which is
pretty undesirable. PackSquash should not make the pack files bigger
than they were before!

Investigating the causes of that size increase yielded the discovery
that color quantization and its dithering were to blame. It usually
happens that quantizing images with lots of dithering improves the file
sizes overall, but this texture was the perfect counterexample for that.

To address this, make PackSquash smarter: instead of blindly using the
color quantized image, check if it yielded any savings compared to just
stripping unnecessary PNG chunks. If so, then use it, but if not, unless
the user really wants quantization, then use the barely optimized
version. As a side effect, images that were already better optimized by
an external PNG optimizer will no longer result in size increase by
PackSquash.

Relatedly, let's also change the default quantization target to a new
special value that signals that not actually quantizing the image is
okay. This way power users can still force quantization to happen if
they really need to.

Some additional refactors to related code were made.