tweak(png_file): bring performance back to the last pre-refactor levels
The refactor we did changed how the number of Zopfli compression
iterations is calculated. This had a dramatic effect in the running time
of our benchmarks, although these increased iterations yielded marginal
time savings. Let's review the Zopfli iteration number calculation logic
to fix some previous calculation errors that made its performance not be
as predictable as expected. While at it, reduce the number of filters
tried in half, which should translate to 50% performance savings too
without much impact on file sizes (most PNG files are well-served by a
few filters, and we are no longer worried on increasing file sizes
compared to the input files).