tweak(png_file): improve error message for PNG files with trailing bytes
The PNG validation and chunk stripping code we introduced for v0.3.1 had
the technically nice side effect of being stricter with trailing bytes
at the end of files, as previously the decoders we used just ignored
them, even if the PNG standard explicitly says that such PNG files are
non-conforming. I expected such an error condition to be very rare, so I
didn't add a specific error message for it.

However, @sya-ri stumbled upon a PNG file that had trailing bytes, and
was puzzled about it, with reason, as the error message pointed out that
the PNG file is "too small". To address that situation, update the wiki
and tweak the error message for future releases.