ci(linux): ditch build-appimage action, use appimage-builder directly
The action was terribly obsolete, to begin with. We updated it to a more
recent version, and it worked, but then we wanted to get rid of some
benign GStreamer warnings, and for this the latest appimage-builder
version needs to run gst-launch-1.0 during build time. So we tried to
put the package that contains that command in the Docker image too, but
then no plugins could be loaded in that environment for some reason.
Using appimage-builder locally, with gst-launch-1.0, worked fine, at
least in a non cross-compilation situation.

We maybe could spend 3 days debugging this and waiting for the Docker
images to upload at a terribly slow speed that reminds me of the dial-up
days, but by the looks of it, not much people cares about the Docker
container anyway, so it's fate is to become obsolete again. Let's
sidestep all of that by not using the action.