chore: rename `packsquash-cli` package. Stop using unstable Cargo feature
By defining a custom binary with the desired name instead of modifying
the filename of the default binary, which is the same as the name of the
package, we can achieve the same result as before, but in a more
straightforward way with less configuration.

While at it, I've discovered that some Rust RFC recommends for words in
crate names to be separated by underscores instead of dashes. This
convention is far from being universally observed, and some prominent
crates use dashes, but we want to do things right, so let's use
underscores.