refactor(windows): revert c6443bf, move back to winres
In the past, we ditched winres because we had the PackSquash library and
CLI binary in the same package, a configuration winres does not support.
However, the CLI binary was moved to its own separate package, rendering
it feasible to use winres again.

winres provided a better, more declarative and efficient way of setting
the resource data of the generated Windows executable, so let's see if
we can start using it again.