fix(pack_file/asset_type): legacy font .bin files can be anywhere
The Minecraft wiki incorrectly pointed out that .bin files belong to
assets/<namespace>/font, but the game interprets the resource location
in an absolute context, so valid resource locations for those files can
point anywhere in the pack. Indeed, a PackSquash user reported dealing
with a pack that had a .bin file outside of the usual folder, which
PackSquash broke due to me coding a too restrictive glob pattern based
on the wiki misinformation.

Make the .bin glob pattern more permissive to fix the situation.