chore(deps): remove transitive dependency on chrono
chrono is a worrying crate to depend on, because it depends on old
versions of the time crate with known security problems. PackSquash is
not affected by these, and overall upstream does not consider the issue
very important, but as GitHub advisories are rolled out for it, silecing
all the security warnings in this convoluted transitive dependency mess
gets cumbersome quickly.

Luckily, chrono is only used by wmi-rs on Windows platforms. The rest of
the ecosystem has been moving away from it, and wmi-rs now has a feature
flag to directly use the time crate instead. Let's enable it to pull
less transitive dependencies and properly fix all these warnings.