chore: add GitHub Codespaces dev container configuration
GitHub Codespaces is very useful for those that have access to it, as it
offers a web VS Code IDE backed by a powerful Linux VM with terminal
access, where development can be done in a breeze, without any local
configuration.

This makes it easier for new contributors to get started, and serves as
a backup development machine for maintainers.