refactor(cli/build): drop vergen dependency
This dependency is relatively heavyweight for what it offers to us, and
it uses a slightly older version of git2. It will probably be updated
soon, but in the meantime that means we can't pull any dependency that
uses the newer git2 version, which is something that we want to do now.
vergen also was a bit troublesome for a contributor that wanted to make
a better PackSquash macOS port.

Let's reinvent a better wheel ourselves here.