fix(audio_file): ignore pitch shift for third pass when not transcoding
Following the trend of the previous commit, it also stood to reason
that, when setting a pitch shift coefficient different than 1 in the
options, it may happen that no pitch shifting is actually done, due to
transcoding being disabled.

Handle that edge case that may restrict the third pass optimization
fallback from happening when there are no adverse side effects to it.