chore: cleanup GStreamer remnants
GStreamer has served us well, but it's time to say goodbye to it for the
greater good of the project. Now we can bend over backwards with other
distribution and build problems, such as universal macOS binaries. Isn't
it great to trade one problem for another, instead of piling them up? Of
course, they can't go away :)