feat: improve error messages for invalid options files
The error messages shown when parsing options files could be pretty
unhelpful for end users when file-specific options are invalid, because
they just pointed to the fact that file-specific options could not be
parsed without any further context or information.

Luckily, I'm not alone this time. Someone made a great PR for upstream
code that has been unmerged for years and makes the messages more
verbose in these cases, so at least it's possible to know what exactly
went wrong. Let's use their great work in PackSquash.

In addition, the path in which the error occurred is now shown, if
available.