chore(deps): update locked dependencies at Cargo.lock
It's necessary to do this for now due to a new Clippy false positive
that is being handled upstream, which causes update workflows to fail.

Also, the serde_derive patch was amended, so the commit the project was
locked to does no longer exist. This amendment fixed some macro hygiene
issues.