fix(gui/updater): parse manifest update date in the format provided by Tauri
For some reason (likely hastily written code), the update manifests
Tauri sends to the frontend represent dates in Rust's time crate
internal OffsetDateTime debug formatting, which does not follow any
standard, so browsers readily can't parse it as a date.

This arguably is a bad upstream decision, but fixing it at the source
will cause compatibility problems, and I doubt that the Tauri team will
find in things like this enough reasons to release v2.

Let's deal with it by parsing dates in such format ourselves as a
fallback.