fix(util/strip_utf8_bom): always consume as many bytes as read to buf
Draining an ArrayVec preemptively empties the original slice even if no
data is read to the buffer, which generates incorrect output that can
only be detected under certain conditions.

To fix that, do not drain the ArrayVec, and remove elements from its
head instead. This is not a particularly efficient operation to do, but
given the extremely small sizes we're dealing with here it doesn't
matter in practice.