fix(gui/frontend): make document title sync with window title properly
The l10n context may asynchronously change the <title> element text
nodes once translated text is available, but a MutationObserver that is
subscribed to only DOM structure changes may not get triggered when this
happens. This is the case sometimes on Linux WebKit, at least.

To fix that, also observe character data mutations.