tweak(gui/frontend): correct credits button behavior, do not reset progress on scroll out
Previously, the credits button could scroll to a strange place because
it tried to scroll to the inner, overflowed credits element, instead of
the intended credits container block, which has its place on the
document flow.

In addition, let's also tweak the threshold for considering the credits
roll in view, and stop resetting its progress when the user naturally
scrolls out of them.