feat: overhaul shader processing code
As described in issue #187, PackSquash does not support shaders that use
preprocessor directives in external declaration position.

To properly address this issue, it is necessary to equip PackSquash with
a GLSL preprocessor that runs before parsing the source GLSL, as the
GLSL specification mandates for proper GLSL compilation. (Building a
GLSL parser that accepts preprocessor directives anywhere, even in the
middle of tokens, is, if not impossible, an exceedingly complex task.)

This commit improves on the situation by adding such preprocessing
capabilities to PackSquash, so that the preprocessing step accepts
directives anywhere that are not seen by the downstream parser. However,
if they appear in external declaration position, it is possible to
inject them back into the AST, in order to preserve them in the
transformed GLSL source. Only directives that are needed by Minecraft
are injected; others are resolved and removed by PackSquash.

In addition, include shaders now need not be full translation units:
they might be standalone statements or expressions, too. Functionality
to prettify shaders was also added, revamping the `minify_shader` option
to a more powerful `shader_source_transformation_strategy`.

These changes allow PackSquash to deal properly with a wider variety of
shaders. The only ones that still trigger failure modes are:

- Shaders which depend on `#moj_import`ed preprocessor variables to be
  syntactically correct.
- Shaders which depend on `#moj_import`ed preprocessor variables to
  build the expected source code, but are syntactically correct
  otherwise. (This can be worked around by setting
  `source_transformation_strategy = 'keep_as_is'` for the affected
  shaders.)
- Shaders that include vertex or fragment shaders, relying on their
  preprocessor variables or syntax to be valid. (This can be partially
  worked around by setting `is_top_level_shader = false` for the
  included shaders, although including vertex or fragment shaders is an
  extreme edge case that we never saw before, and arguably a bad idea.)

The first two kinds of shaders should be able to be properly parsed for
PackSquash v0.4.0, because its design allows pack files to resolve
references to other files.