perf(png_file): use experimental OxiPNG raw API
This new API renders it unnecessary for PackSquash to encode
intermediate PNGs for them to be consumed by OxiPNG, which is noticeably
faster and more memory efficient, especially for smaller images, where
the encoding overhead is significant.

While at it, the image processing code was refactored to be much more
ergonomic and easier to read, 0 is now rejected as a value for the
`maximum_width_and_height` option, the OxiPNG options were slightly
tweaked to use a better set of filters, and images no longer are
downsized if their color type cannot be changed.