test(gui): add initial automated e2e tests
These tests are meant to be automated smoke acceptance tests, aiming to
guarantee that no build can be utterly broken from an end-user
perspective without a maintainer noticing.

These tests were implemented in a rather elegant way by leveraging the
W3C WebDriver implementations of the OS webviews Tauri relies on, thanks
to WebdriverIO. The acceptance tests themselves are written in Gherkin,
which is considered best-practice for tests so close to the user domain,
and part of the behavior-driven development (BDD) paradigm.