fix(png_file): drop trailing bytes check
This check may cause lots of trouble to unsuspecting, not-so-technical
users with lots of PNG files, all for the little benefit of complying to
a standard that few applications are sensitive to anyway. Shame on image
editors for generating this garbage, but we should handle it better,
while letting users know that something is wrong.

Future releases will turn this to a warning. This can't be done now
because the current design does not allow this.