refactor: drop `open_files_limit` option
This option is rather advanced to use, and so far I have not seen a
legitimate reason for users to change its default value. The only times
it's been included in an options file is because people copied the
complete example on the wiki, despite my advice to the contrary.

To prevent people from shooting themselves in the foot while
accomodating advanced needs, let's drop the option and make PackSquash
try to raise the file descriptor limit as needed for the desired
concurrency level. If that fails, fall back to throttling concurrency
and output an user-friendly warning explaining what happened and how to
get the intended performance.

A convenient way to test these changes is by using a command like this
on a Linux workstation:

$ systemd-run --user -GPdt -p LimitNOFILE=<soft>:<hard> target/release/packsquash /tmp/packsquash.toml