refactor: remove global audio options
These were a bit weird and complicated, and the file-specific options
mechanism is powerful enough to implement advanced use cases, even
though it might be verbose.

While at it, let's assert that no error happens when creating a
resampler, as such a thing signals an internal logic error worth of
panicking.