fix(audio_file): prevent resampling buffer sizing errors and artifacts
The Discord user .furq reported that PackSquash did not deal with mono
audio files sampled at 32768 Hz properly. In particular, the first issue
that arised was that Rubato was using an improper output buffer size
calculation, which was reported and is in the process of being fixed
upstream.

After this fixed was patched in PackSquash, I decided to listen to the
resulting resampled file and noticed that its output sounded wrong, with
lots of unexpected filler silence. It was a bit of an involved issue to
locate, but I traced it to the resampler outputting empty audio buffers
when it needs to buffer more input samples and libvorbis deciding to
output silence even when the input buffers are empty.

.furq has kindly allowed us to incorporate their technically interesting
file to our test suite. Thank you!