tweak: disable `downsize_if_single_color` by default
While a good idea in theory, the experience with v0.4.0 has shown that
more packs are broken by this optimization than helped. Let's disable it
by default for now, but with a disposition to apply it in a smarter way
in the future, once the necessary refactors to detect how the image is
used by the game are in place.