feat(wip): begin PackSquash redesign
Let's start from scratch and model the problem using totally different
internal abstractions, which have already proven to be helpful in this
proof of concept, WIP code for optimizing blockstates pretty well.

There are lots of things to do, document and review. This is by no means
a complete PackSquash build.