feat(gui): gracefully handle Tauri runtime run errors
A Tauri runtime may fail to create a webview for the application window
if the necessary OS webview components are missing. In general, it may
also fail to launch the application window for other reasons.

An application can't do much to recover from these situations, as they
signal environment configuration errors or logic errors that must be
fixed, usually. But the previous behavior of panicking when they
happened made troubleshooting them harder than necessary on Windows,
where getting the stderr output of an application that does not have
attached consoles is, frankly, out of reach for most of the GUI target
audience, and thus a potential source of support inquiries.

Let's handle this error better by showing it on a message box that
should help both newbies and power users alike.