chore(gui): fix unconforming use of "auto" for SVG geometry properties
The SVG 1.1 spec does not document "auto" as a valid value for the
"width" or "height" <svg> element attributes, which must be a CSS
length:
https://www.w3.org/TR/SVG11/struct.html#SVGElementWidthAttribute. This
specification states that if the attributes are not specified user agents
should behave as if a 100% length was set.

SVG 2 relaxed this requirement, allowing the "width" and "height"
properties to be specified via CSS, and take the value "auto", which is
to be interpreted as the relative length 100%.

However, no browser supports the full SVG 2 specification, and thus the
handling of "auto" depends on the browser: some use a hardcoded length,
while others default to 100%, but all of them treat such a value as an
error that gets printed to the console.

Thus, there are two ways to get the behavior we want without triggering
undefined behavior in common browsers:

- Set the "width" and "height" attributes to "100%" (note that
  phosphor-icons requires always setting these attributes).
- Set the "width" and "height" CSS properties to "100%" (CSS properties
  override the respective <svg> element attributes).

Let's implement those two ways, choosing between them as to what causes
less overrides to be made, and thus more maintainable layout.