refactor(properties_file): simplify, standarize on Windows-1252 encoding
The Java documentation specifies that .properties files should use the
Windows-1252 (also imprecisely known as ISO-8859-1) encoding, but there
was code for supporting the Unicode BOM, which should not appear on the
first place and does not help with incompatible misinterpretations of
the source encoding as UTF-8; it'd only help if a UTF-8 with BOM file
only uses ASCII characters, but break in other cases.

In the future, we might study and implement ways of handling several
encodings, depending on the encoding expected by the mod using these
files.