fix(shader_file): fallback to compat. no transform if `#moj_import` is involved
As highlighted in issue #187, a significant number of practical shaders
use `#moj_import`s in ways that are tricky for PackSquash to handle
without import expansion capabilities. The necessary redesign to achieve
expansion will ikely take some time, but there is stakeholder pressure
to get better (and more correct) solutions working now: PackSquash has
arguably been incorrect in how it processes shaders for months, and
justifying the status quo with upcoming plans to tackle the problems at
their source is no longer a tenable proposition, leading to
relatively frequent uncomfortable situations for both me and users.

Therefore, let's play the cards very close to our chest when it comes to
transforming GLSL sources: don't do it unless we can be certain that we
have a full AST and all the preprocessor state is known. This entails
significant optimization regressions for any shader that uses
`#moj_import`, as not expanding that directive is the root of all evil,
but PackSquash will at least just work in much more cases. In
particular, these changes were tested with FancyPants, TextEffects v2.1
and objmc, which are interesting shaders from a parsing standpoint due
to their popularity and intensive usage of GLSL language features.