ci: bring back GitHub Actions cache for Windows
According to all known laws of informatics, sccache should work well on
Windows. But for some reason in our environment Windows builds with the
MinGW toolchain get so slow due to sccache overheads, even though our
current sccache approach with a Redis database storing compilation
aritfacts is overall a net improvement for the remaining targets. So
let's fall back to the older cache method for Windows to continue
cutting down on CI time.