docs(README): remove stale JetBrains FOSS development license notice
JetBrains has been running a FOSS development license program for years
now, by which they give away yearly licenses for their products. At some
point after March 2024 and before February 2025, they quietly changed
the conditions to remove a list of project requirements in favor of a much
more vague partnership program that looks for projects with communities
that make "valuable contributions to the software industry" (in my
opinion, that's marketing speech for projects that give them good PR, or
that integrate some part of the typical tech stacks big companies use).

My FOSS development license expires on March 8th, so I requested a
renewal on February 24th as prompted by automated emails and as I've
done in previous years. I have not yet received a response in 9 business
days, I'm not aware of any circumstances that may delay a reply to this
request, and in previous years the process has been much faster despite
express warnings of potential delays in processing these requests. March
8th is Saturday, and I don't expect such a non-urgent request to be
processed on a weekend, when most employees in such a company are not
working. Therefore, it seems fair to conclude that JetBrains is not
interested in renewing a FOSS development license for PackSquash.

As a result, the related acknowledgment has become false, and should
either be modified or removed. As I'm unaware of the copyright terms
under which the JetBrains logo can be used (JetBrains has stated "feel
free to use the JetBrains logo" on their license request approval
emails, but that does not mean I'm still free to use it without some
sort of approval from JetBrains), and I don't have particularly strong
reasons to be appreciative of the fact that they effectively pulled the
rug out from under the project and didn't even bother with a canned
negative reply, let's remove it. (To be fair, they started offering
non-commercial subscriptions for a subset of products, and the employees
I've interacted with in past years were all nice and professional, but
this sort of changes tend to come from cold execs higher in the chain,
so it's fair they also get a cold response that highlights with simple
numbers that they might not be making the best decisions for truly
independent and original projects like PackSquash.)