feat(squash_zip): support previous ZIPs with comment strings
When adding arbitrary comment strings to a ZIP file, the end of central
directory (EOCD) record stops being at a constant offset from the file
end. In upcoming commits we'll make this comment string configurable, so
for ZIP file reuse to keep working in that scenario we have to implement
a more flexible EOCD locating algorithm that drops that constraint.

While at it, I've refactored some things.