feat: migrate redirections from Firebase Dynamic Links
As expected with almost every nice Google product, Firebase Dynamic
Links will be sunset for good on August 2025. Now that we have a pretty
domain and a website linked to it, let's use it for redirects that are
hopefully stable for many years to come. (It's amazing how controlling
your own domain, which costs money, is still the most reliable solution
for setting up some simple redirects.)