tweak(ci): improve maintainability of Debian package desc generation code
`html2text` is a tool whose maintenance status is quite uncertain, and
has introduced breaking changes in its interface that will bite us as
soon as we dare to update the Debian container distribution version.

So, instead of stitching together sed scripts with obsolete CLI tools,
let's just converge on using the `markdown-it` Python port from a small
Python script, which should be much more robust and flexible moving
forward, as the Markdown is now parsed and rendered directly to plain
text by a proper and well-maintained library to the effect. Providing
Python 3 is already available on the environment, this even slightly
reduces the total size of Debian dependency packages needed for the
script to work.